Mac: Re-enable CodeQL for main build
We disabled CodeQL for the main build (and enabled it for the legacy
build) as part of #2564. We did that as the signed release build broke
with CodeQL enabled. One likely reason was #2925, which was
only found later.
In addition, it seems like the codesign invocation has to run with
unmodified linker environment flags.

Fixes #2926